Moves signup data to own data file, sets up create user request
