+++ frontend/src/pages/signup_volunteer.jsx
@@ -15,6 +15,9 @@ import {
 } from '@material-ui/core';
 import LocationOnRoundedIcon from '@material-ui/icons/LocationOnRounded';
 
+import api from '../helpers/api';
+import { checkboxSections, disabledStaff } from '../data/signup';
+
 const useStyles = makeStyles((theme) => ({
   heading: {
     marginTop: theme.spacing(8),
@@ -31,9 +34,32 @@ const useStyles = makeStyles((theme) => ({
   },
 }));
 
-function handleFormSubmit(fields, form) {
+async function handleFormSubmit(fields, form, history) {
   console.log(fields);
   form.setSubmitting(false);
+
+  // TODO - check for errors
+
+  const payload = {
+    user_email: `test${Math.random()}@test.com`,
+    user_first: 'first',
+    user_last: 'last',
+    user_password: 'password',
+    user_is_active: true,
+    user_is_admin: false,
+    user_is_medical_professional: false,
+    user_is_verified: false,
+    user_is_volunteer: true,
+  };
+
+  // TODO - handle error response; handle success response
+
+  const response = await api.createUser(payload);
+  console.log('signup response', response);
+
+  if (response.user_id) {
+    history.push('profile');
+  }
 }
 
 function StaffCheckbox({ staff, arrayHelpers, disabled }) {
@@ -59,53 +85,9 @@ StaffCheckbox.propTypes = {
   disabled: PropTypes.bool.isRequired,
 };
 
-function SignUpPage() {
+function SignUpPage(props) {
+  const { history } = props;
   const classes = useStyles();
-
-  const staff = {
-    cad: [
-      { label: 'Solid Works', id: 'solid-works' },
-      { label: 'Inventor', id: 'inventor' },
-    ],
-    medical: [
-      { label: 'Registered Nurse', id: 'registered-nurse' },
-      { label: 'Physician Assistant', id: 'physician-assistant' },
-      { label: 'Medical Student', id: 'medical-student' },
-      { label: 'Intern', id: 'intern' },
-      { label: 'Resident', id: 'resident' },
-      { label: 'Attending', id: 'attending' },
-    ],
-    engineering: [
-      { label: 'Mechanical Engineer', id: 'mechanical-engineer' },
-      {
-        label: 'Mechanical Engineer with FEA experience',
-        id: 'mechanical-engineer-with-fea-experience',
-      },
-      { label: 'Electrical Engineer', id: 'electrical-engineer' },
-      { label: 'Mechatronics Engineer', id: 'mechatronics-engineer' },
-    ],
-    legal: [
-      { label: 'Lawyer', id: 'lawyer' },
-      { label: 'Barrister', id: 'barrister' },
-      { label: 'Paralegal', id: 'paralegal' },
-    ],
-    manufacturing: [
-      { label: 'FDM 3D printer', id: 'fdm-3-d-printer' },
-      { label: 'SLA 3D printer', id: 'sla-3-d-printer' },
-      { label: 'SLS Nylon 3D printer', id: 'sls-nylon-3-d-printer' },
-      { label: 'Machinist', id: 'machinist' },
-    ],
-  };
-
-  const checkboxSections = [
-    { staff: staff.medical, label: 'Medical Staff Advisors' },
-    { staff: staff.engineering, label: 'Engineering' },
-    { staff: staff.manufacturing, label: 'Manufacturing' },
-    { staff: staff.legal, label: 'Legal' },
-    { staff: staff.cad, label: 'Computer-Aided Design' },
-  ];
-
-  const disabledStaff = ['lawyer', 'electrical-engineer', 'intern', 'inventor'];
 
   return (
     <Formik
@@ -115,9 +97,9 @@ function SignUpPage() {
         selectedField: 'medical',
         soughtStaff: [],
       }}
-      onSubmit={handleFormSubmit}
+      onSubmit={(fields, form) => handleFormSubmit(fields, form, history)}
     >
-      <Form class={classes.form} noValidate autoComplete="off">
+      <Form className={classes.form} noValidate autoComplete="off">
         <Grid justify="center" container>
           <Grid item xs={11} sm={11} md={11} lg={9}>
             <Grid container>
@@ -264,5 +246,15 @@ function SignUpPage() {
     </Formik>
   );
 }
+
+SignUpPage.propTypes = {
+  history: PropTypes.shape({
+    push: PropTypes.func.isRequired,
+  }),
+};
+
+SignUpPage.defaultProps = {
+  history: {},
+};
 
 export default SignUpPage;
